+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## CHANGES
 
+ - [BC-BREAK] Remove `Registry::setDelegate`.
  - [BC-BREAK] Default values in parameters are not resolved.
  - [BC-BREAK] Given `$make` function doesn't retrieve dependencies, just make object without cache.
  - New `Registry::freeze` to disallow new values or factories to be assigned.